Move components into feature-specific directory
